+++ packages/elements/src/types/clerk.d.ts
@@ -1,9 +1,0 @@
-import '@clerk/shared/types';
-
-import type { EnvironmentResource } from '@clerk/shared/types';
-
-declare module '@clerk/shared/types' {
-  export interface Clerk {
-    __unstable__environment: EnvironmentResource | null | undefined;
-  }
-}